bit more refactoring
