+++ kafka-client/src/test/java/dev/responsive/integration/ResponsivePartitionedStoreEosIntegrationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.responsive.internal.stores;
+package dev.responsive.integration;
 
 import static dev.responsive.testutils.IntegrationTestUtils.createTopicsAndWait;
 import static dev.responsive.testutils.IntegrationTestUtils.getCassandraValidName;
